chore: rename streams file FUI-1754
